fix(axi,config): use Mux1H for r.ready routing; resetVector to BigInt

+++ core/src/main/scala/markorv/config/Config.scala
@@ -132,7 +132,7 @@ case class PmaConfig(
 
 case class CoreConfig(
     simulate: Boolean,
-    resetVector: Int,
+    resetVector: BigInt,
     fetchQueueSize: Int,
     asidWidth: Int,
     axiConfig: AxiConfig,
